unix: put clang toolchain on a diet
Our goal of building a Clang toolchain is not to provide a generally
usable Clang toolchain: our goal is to build Python and its
dependencies.

With that in mind, this commit puts the Clang toolchain on a diet.

We stop building libcxx and libcxxabi because we're using the libgcc
variation at runtime anyway.

We stop building compiler-rt and ldd until stage 3 on Linux because
neither is used as part of stage 1 or 2.

We drop fuzzers, sanitizers, and other misc pieces from compiler-rt
because the only thing we likely care about is the PGO profiling
support.

On macOS, we limit targets to aarch64 and x86, as those are the only
architectures supported by Apple operating systems.