unix: more clang toolchain dieting
We disable some unused features to speed up the build.

compiler-rt builtins are disabled because we don't use the compiler-rt
runtime for anything except PGO.

We also only emit the x86 target as that is the only target currently in
use by this project. (Cross compiling currently uses a different
toolchain, unfortunately.)

Altogether, this reduces the size of the tarball by ~200 MB and should
speed up builds in CI.