unix: write out a file with target settings
Previously, if we changed some values in `targets.yml` the dependencies
in the Makefile wouldn't be invalidated and we wouldn't rebuild packages
when critical settings changed.

This commit fixes that by writing out a file with target-specific
metadata derived from `targets.yml`. This new file is a dependency of
all non-toolchain targets, ensuring that any meaningful changes to the
target settings result in a rebuild of all non-toolchain dependencies.