unix: build targets in parallel
I'm pretty confident the Makefile dependencies are reasonably correct
these days and we can actually build projects in parallel. This wasn't
always the case. And combined with lack of output to help debug
failures, is likely the reason this was never enabled.

We now prefix output by the entity being built. And we save logs of
each component to its own file. So I don't see a good reason to not
leverage parallelism during builds.

On my machine, this change speeds up builds by ~1 minute. Hopefully it
results in a nice speedup in CI as well.