unix: don't export dependency symbols on macOS
This commit follows up on recent work we did for Linux/ELF around
not exporting dependency symbols and applies it to Apple/Mach-O.

We change compiler flags so symbols are hidden by default. We also
teach the linker to make symbols in library dependencies hidden
by default to prevent them from being exported / dynamic in the
resulting Mach-O binary.

Mach-O linker support for controlling symbol visibility is different
from ELF. We had to utilize an `ld64` specific argument
(`-hidden-l<lib>`) to force symbol visibility on a per library basis.

The Rust validation code for Mach-O has been updated to verify we
aren't exporting dependency symbols and are properly exporting Python
symbols, just like we did for ELF. Distributions built before this
fail the new validation.

This should finish the implementation of #114. Windows is still not
addressed. We can file a separate issue to track Windows.