unix: patch configure.ac instead of configure
Python 3.11 introduces a configure based mechanism for configuring
extension modules. It is looking like we'll need to bulk modify
behavior to get it to work with our opinionated way of statically
linking extension modules.

In preparation for that, this commit changes all our configure patches
to patch `configure.ac` instead of `configure` so we can make changes at
the source without bloating patch size. As part of this, I also moved
the patches to standalone files, as having to copy inline patches with
tabs was historically annoying. This approach is much easier.