unix: install m4 and autoconf
A future commit will introduce patches to configure.ac, which means we
need to run autoconf. As macOS doesn't provide a sufficient autoconf
and we want to use the same version of autoconf everywhere, it makes
sense to build autoconf from source for all targets.

So this commit does that.

Autoconf requires m4, so we need to build and install m4 as well.