unix: replace disabled-static-modules files with YAML
I want to move to using YAML to describe the extension module
configuration. The easiest place to start is with the set of disabled
modules.

This commits effectively converts all the metadata in
disabled-static-modules-* files to YAML.

As part of this, the `_gdbm` extension module may have been disabled
on a few target triples not previously disabled. I think that's fine.