unix: unify container image building
We add support for building just the container images in
`build-linux.py`. In CI, we add jobs for building the images.
We then save images as artifacts and download and load them in
individual jobs. This should make CI much faster as we avoid
redundantly building images on every target build.