unix: always link against libedit
Before, CPython 3.8 and 3.9 on Linux linked against readline by default
and built a variant that linked against libedit. After, all CPython
versions on POSIX platforms link against libedit and don't publish a
readline variant.

In my opinion maintaining the code to support extension module variants
is too much complexity. If my memories are correct, the reason I
introduced this complexity was to avoid taking a ton of patches to
CPython. This was at a time where we had significantly fewer patches and
I was trying to keep the CPython source as close as possible to
upstream. Fast forward a few years and we have a ton of patches and I'm
no longer scared to take patches. Another factor was likely that
readline.c required a massive patch to support libedit outside of macOS
because of excessive use of `#ifdef __APPLE__`. This code was cleaned up
in Python 3.8 (I think) and patches to support libedit because much more
manageable.

This commit removes the readline extension module variants, changes
libedit to be the default library, and deletes references to readline
from the code base. Good riddance.