unix: port _asyncio definition to YAML
Now that we have static-modules deleted, we can continue to subsume
content from the Modules/Setup file and do away with some of our
hacky parsing of that file to preserve commented lines.

This commit starts the process by defining the `_asyncio` module in
YAML and undoing the dependence on the commented entry copying. The
plan is to do this for all extension modules using this feature.