pythonbuild: remove extension module variants
We formerly used variants to build multiple versions of an extension
module. e.g. a readline extension linked against either libreadline or
libedit.

In recent days we've removed all use of variants. The code to support
variants is somewhat complex and in my opinion not worth keeping around
because of this complexity.

This commit deletes all code for handling variants.

Note that PYTHON.json still advertises variants. I may or may not
change the JSON schema to remove variant support. For now, it is
easier to avoid the backwards compatibility break.