unix: link _decimal against standalone libmpdec
The previous commit enabled support for building mpdecimal as a
standalone project. This commit replaces the inline build of
mpdecimal as part of CPython to linking against our standalone
mpdecimal.

On Python 3.8 and 3.9, this effectively upgrades mpdecimal to 2.5.1.
This required a minimal patch.

As part of this, we also fix the defines on macOS. UNIVERSAL is no
longer needed, as this is handled by mpdecimal's build system. But
_decimal.c does need one of the CONFIG_ defines set or else things
blow up.