unix: use extensions metadata to derive PYTHON.json build info
Before, we fed the `Setup` and `config.c.in` [parsed] contents into
`python_build_info()` to derive PYTHON.json metadata.

The recent refactorings to define all extension metadata in YAML
and to ensure it is in sync with distribution reality enables us to
greatly simplify matters.

After this commit, PYTHON.json metadata is now derived from data
structures returned by `derive_setup_local()`, not the raw Setup
files or config.c.in content directly.

A side-effect of this change is that metadata looks more reasonable
now. diffoscope says that recent refactorings had subtly broken
advertisement of the `required` field. diffoscope also tells me
that at least the `_weakref` extension was advertising itself twice.
This is probably due to picking up a source from both config.c.in
and Setup. Now that we have a single dict of enabled extension
modules, this duplication is removed. It appears this duplication
pre-dated the most recent release and the refactorings that have
come before.

There's still room to improve things. For example, we could derive all
the metadata in `derive_setup_local()` and not parse Setup lines in
`python_build_info()`. But this is the domain a follow-up.