unix: add and use environment variables for Python version tests
As we add Python 3.11 support, the number of places where we need to
explicitly check for 3.11 highlights how poor our version comparison
shell logic is.

This commit introduces and uses new environment variables that
advertise whether Python meets a minimum and maximum version. This
allows us to express most version compares in terms of a version
range instead of a direct compare. This makes it easier to add new
versions without having to update every version compare call site.