unix: refresh makesetup patch to remove shared libraries
`makesetup` source code changes in 3.11 invalidating this patch. We
could potentially drop this patch completely by disabling the extension
via configure. But preserving the existing patch seems harmless.

As part of this we also dropped the deduplication of `OJBS`. I want
to say that this deduplication was necessary because prior versions of
CPython included the same source/object file in multiple extensions.
e.g. one of the math sources. I don't think we are seeing this in 3.11
so it should be safe to drop.