unix: nerf PY_STDLIB_MOD maco on Python 3.11
See the inline comment for details.

3.11 made strides to move extension module configuring from setup.py
to configure. This functionality looks promising. But there is still
a fair amount of logic for some extensions in setup.py. So it isn't
fully usable yet.

Furthermore, configure functionality appears to still not support
statically linking extensions.

Rather than try to make the new functionality work, this commit adds
a patch to simply disable it. We still generate a Setup.local file
which still has the intended effect of statically building extensions.
setup.py is still mostly nerfed to a no-op.