windows: suppress duplicate symbols warnings in static builds
Our hacked up static builds include various libraries multiple times.
This sets off a tsunami of linker warnings for duplicate symbols.

This behavior has existed for ages and the best I can tell the
linker sorts it out without issue. So this commit suppresses the
warning.

The build logs are now vastly easier to read as a result.