rust: add validation that libpython exports extension initialization functions
The recent oversight with pwd in 3.11 gave me the idea that we could
also validate that referenced initialization functions are also exported
from libpython. This is important since our metadata + code derives the
initialization function name and there's the possibility for these to
get out of sync. That being said, the linker should catch some obvious
errors since these functions are referenced in config.c. But the linker
can't detect omissions. So validation is still warranted.

While implementing this, I discovered that we're currently exporting
these initialization symbols on Linux and macOS and official Python
builds aren't. We may want to change this someday to better conform
with Python's ABI.