rust: capture libpython exported symbols on Windows
We weren't doing this before and some of our validations weren't
working on Windows as a result.

This change enables us to apply the exported symbols verification
to all distributions. In order to do this, we did need to exclude
static distributions (Linux musl and Windows static) from the check.