downloads: upgrade binutils 2.38 -> 2.39
2.40 was released hours ago. But I don't see the .tar.xz on ftp.gnu.org
yet. So we stick to 2.39.

Reason I'm doing this is I noticed that our invocations of `objdump`
when building Python were emitting `Unrecognized form: 0x22` warnings.
This is due to DWARF emitted by LLVM 15 being too new for the previous
version of binutils. This was likely harmless. But modernizing binutils
shouldn't hurt.