unix: enable _testinternalcapi
This effectively reverts commit
5611869908039ff983f77359a14450e902ccf1a3.

The lack of this extension was undermining PGO.

Enabling this on 3.11 was non-trivial because its bootstrap Python
didn't want to link with this extension, which is typically built
shared. It was easier to always build a host Python than to coerce
CPython's build system to have its bootstrap Python work with this
extension.

AFAICT Python < 3.11 doesn't have any bootstrap issues with enabling
this module.