unix: always build and use host Python when building 3.11
See inline comment.

I'm running into a lot of pain with `_bootstrap_python` and
`Programs/_freeze_module`. CPython's build system makes a lot
of assumptions that certain modules / symbols won't be in the
bootstrap Python. We invalidate that by statically linking
everything.

The path of least resistance is to build and use a host Python for
freezing as it avoids building the half-complete binaries altogether.