unix: build _crypt extension module as shared
Modern Linux distributions are starting to remove `libcrypt.so.1` from the
base disto. See #173 and #113 before it for more context.

The `_crypt` extension module depends on `libcrypt.so.1` and our static
linking of this extension is causing the full Python distribution to
depend on `libcrypt.so.1`, causing our binaries to not load/run on
these distributions.

This commit adds support for building extension modules as shared
libraries (the way CPython does things by default). We update our
YAML config to build `_crypt` as a shared library.

As part of this we remove a patch to `makesetup` that was removing
rules for emitting extension info to the generated `Makefile`. Git
commit messages and my mental memory are a bit foggy on the history.
But I want to say this patch had its origins in the very early days
of this project when I was still trying to figure out how to get
extension modules to statically link. The removed code created some
contamination of the generated `Makefile` making it harder to grok
behavior. I think that the presence of these additional variables and
rules is safe and doesn't cause unwanted regressions. We have pretty
strong validation of the built distributions - including ELF symbol
visibility checks. So I'm optimistic this reintroduction won't cause
problems.

Closes #173.