Build OpenSSL 3.0 without atomic ops when not available
We need to tell OpenSSL 3.0 to not use atomic operations on
architectures without native support for them - in this instance,
OpenSSL will use a fallback implementation with locks.
The other option is to provide a software implementation of the atomic
operations with libatomic, but I found this difficult to get linked
correctly on all platforms.