downloads: CPython 3.11.4 -> 3.11.5
We remove a patch for limits.h handling because upstream incorporated a
fix (from a bug report I filed).

And it looks like upstream backported build system changes to build with
OpenSSL 3.0 on Windows. We need to update our code accordingly.

We really should be building against OpenSSL 3.0. But that's scope
bloat. We'll get there.