unix: add support for and use OpenSSL 3.0.10
Now that the OpenSSL version is parameterized we can add OpenSSL 3.x
build support with little effort.

The OpenSSL 3 license changed to Apache 2.0. And its build system would
default to `lib64` instead of `lib` for the library directory in some
configurations. And the set of available hashing algorithms changes.

Probably the most annoying change here was dealing with libatomic. It
looks like OpenSSL 3 links against libatomic in many scenarios. Clang
ships a dynamic libatomic.so by default. So if we're not careful we
get a run-time dependency on libatomic.so. We work around this by
using the `-l:libatomic.a` linker flag syntax when building/linking
CPython extensions using libcrypto to force the symbols to get linked
statically.

We choose to use OpenSSL 3.0 even though 3.1 is available because
3.0 is an LTS release and has a longer support window. And CPython
right now appears to be using 3.0 more than 3.1. While 3.1 appears to work
(I initially used 3.1 before downgrading to 3.0), OpenSSL compatibility
has historically had quirks with CPython and I don't want to venture
too far into uncharted territory with 3.1.

We still need Windows support. But at least macOS and Linux are out of
the way.o

Part of #178.