global: CPython 3.12 support
I cobbled this together over several partial days of work. There's a lot
of changes in there and I'm too lazy to split it back up into microcomments.

I haven't performed a full audit of the produced archives and their
PYTHON.json files. I suspect there's some subtle bugs lingering in here:
there usually is for the initial release of a major version. But
CI and our Rust validation is passing and that usually means a lot.
So I'm going to commit what we have and iterate on it as needed.

Closes #186.