unix: build host CPython separately
Before, we built a host CPython inside build-cpython.sh for build
configurations that required it.

This commit extracts the host CPython build to its own build script and
make target.

The impetus for this is that iterating on changes to the main CPython
build script is slow due to having to recompile the host CPython all
the time. I've been wanting to make this change for a while to speed
up that common development workflow.

As part of this refactor we now sometimes compile the host CPython
when we don't need to. But since we always the host CPython for
3.11+, eventually that waste is reduced to 0. So I'm fine with the
overhead.