unix: enable BOLT optimizations on Linux x86-64
BOLT appears to only work on Linux. And it only supports x86-64 and
aarch64. And since we don't have PGO builds for aarch64, that means we
can only ship it on x86-64. (We can't even enable support in the build
system since we don't have a sufficiently modern Clang to use on aarch64.)

In order to get this to work, I had to force DWARF version 4 to avoid
BOLT crashing. This is related to #155. I plan to file an upstream issue
on the LLVM crash.