pythonbuild: extract tarballs as build user
I'm not sure why we were forcing use of the root user here. Maybe a
holdover from prior code?

Writing the files as the build user enables jobs to modify their
content, which can be desirable for patching these files in individual
builds.

Impetus for this is I want to add a `cc` symlink for `clang` so the
compiler gets reported as `cc` and not `clang`.

This change only impacts container builds. Since those are reproducible,
I want to say this is safe if it passes CI.