unix: normalize build configuration to aid portability
This commit implements a long desired feature to normalize the build
configuration in various distribution files post build but pre
packaging.

The goal of this general feature is to make distributions highly
portable. Before, configurations (which were used to e.g. compile
extension modules) referenced build environment paths, like `/tools`.
This is not desirable and can confuse downstream users when unexpected
settings are used.

The impetus for this work is #194. As part of this change we strip the
`-fdebug-default-version` argument from `CFLAGS` to restore CFLAGS
compatibility with GCC.

There's no doubt additional settings that could be normalized. Those
can be implemented as follow-ups.