ci: use upload-artifact@v4 for release publishing
We can no longer upload to the same named artifact across
multiple jobs. So we need to incorporate the build optimization level
into the artifact name.

I think release automation will continue to work since we iterate
over every artifact, download, and inspect the files within. But there
is a chance we need to tweak release automation to account for this
change.