downloads: upgrade LLVM toolchains
The Linux toolchain should be unchanged.

The Apple toolchains now contain LLD and should use it by default.

Previously, we would use `/usr/bin/ld` on macOS machines. In GitHub
Actions, the system linker has a fixed-in-LLVM-15 bug where weak symbols
are not preserved during LTO. This was causing LTO'd binaries to
strongly reference symbols from too-new Apple SDKs and for binaries
to fail to run on those OS versions.

This bug has apparently been present for ages. But we failed to realize
the full extent of it until we attempted to upgrade macOS GitHub
Actions runners and encountered CI failures when verifying symbol
macOS version compatibility against metadata in the Apple SDK. Yay
for CI coverage of that!

Closes #216.