rust: upgrade octocrab
We had to bypass octocrab for release asset upload because octocrab's
generic HTTP API forces us to use String, which binary assets are not.
This is tracked upstream by https://github.com/XAMPPRocky/octocrab/issues/360
(an issue I filed last year).

I haven't tested the new code and there's a chance we broke uploads. If
so, we can revert this commit.