unix: rework terminfo database handling
Before, ncurses on Linux was built to look for the terminfo database in
a location that only existed in the build environment

After, it looks for the database in `/etc/terminfo`, `/lib/terminfo`,
and `/usr/share/terminfo`. These are the directory locations that
work on a superset of Debian and RedHat. This hopefully means that
terminfo database resolution "just works" in a lot of common Linux
environments.

As part of this, we add the terminfo database to the ncurses package
artifacts. Then we copy the terminfo database into
`install/share/terminfo` in the CPython distribution if a terminfo
database is present. macOS doesn't distribute a terminfo database
because the system terminfo database should be "good enough." Plus
we don't build ncurses on macOS distributions.

Closes #215.