downloads: switch URL for xz; add comment about backdoor
Due to the discovered backdoor in xz, GitHub took the official
project offline. The old download link was apparently redirecting
to GitHub and is now broken.

I'm guessing this will resolve in time. But for now let's work
around by using a "self-hosted" version of the archive.

Yay for content digest pinning to prove this is exactly the same
copy of the source code.

Closes #242.