Patch `_tkinter.c` to look in base prefix (#421)
## Summary

This PR attempts to adresss one of the [known quirks](https://gregoryszorc.com/docs/python-build-standalone/main/quirks.html#tcl-tk-support-files) around Tcl/tk files. On Windows, CPython will look for the Tcl/tk files at `{base_prefix}/lib`. That works well for python-build-standalone, since we ship the files in that location. However, on Unix, CPython does _not_ check that location -- instead, for python-build-standalone, it just looks relative to the executable. This works for non-virtual environments, since CPython _will_ look in `{prefix}/lib`. But not for virtual environments.

The patch applied here ports the Windows discovery logic to Unix, so that we always hook up the right paths.

The advantage of this approach (vis-a-vis others that were considered) is that it doesn't rely on injecting an environment variable into the user's environment.

For now, I've only patched Python 3.13. Unfortunately, I'll need a _slightly_ different patch for each Python minor. I've already authored them locally, but I want to hold off on wiring them all up until we're aligned on this general approach.

In theory, I think this change could be upstreamed.

## Test Plan

- Build Python locally.
- `./python/install/bin/python -c "from tkinter import Tk; Tk()"`
- `./python/install/bin/python -m venv .venv`
- `.venv/bin/python  -c "from tkinter import Tk; Tk()"`