Convert to UTF-8 prior to setting Tkinter path (#425)
## Summary

In https://github.com/indygreg/python-build-standalone/pull/421, @carljm
pointed out that `path` was uninitialized in `setenv("TCL_LIBRARY",
path, 1);`. It turns out we need to set it via `PyUnicode_AsUTF8`.

I confirmed that the existing code _was_ running; it's just that`setenv`
returns an error (rather than crashing), and we weren't checking the
result (though the same is true in the Windows path). If you, e.g., try
to `printf` `path` just before, you get a segmentation fault.

This code isn't necessary to fix the motivating Tkinter limitation, so
that's another reason it wasn't caught.